Use CachedValue as base class
